// src/common/enum/audit_log.enum.ts
export enum AuditLogAction {
  CREATE = 'CREATE',
  UPDATE = 'UPDATE',
  DELETE = 'DELETE',
  LOGIN = 'LOGIN',
  LOGOUT = 'LOGOUT',
  PASSWORD_RESET = 'PASSWORD_RESET',
  ACTIVATE = 'ACTIVATE',
  DEACTIVATE = 'DEACTIVATE',
  ARCHIVE = 'ARCHIVE',
}

export enum AuditLogEntity {
  USER = 'USER',
  PRODUCT = 'PRODUCT',
  SUPPLIER = 'SUPPLIER',
  STOCK = 'STOCK',
  PURCHASE_ORDER = 'PURCHASE_ORDER',
  PRODUCT_SOURCE = 'PRODUCT_SOURCE',
  CATEGORIES = 'CATEGORIES',
  BUSINESS = 'BUSINESS',
  STORE = 'STORE',
  STOCK_MOVEMENT = 'STOCK_MOVEMENT',
}